requirements.txt
Django>=2.2
celery==5.2.7
redis
djangorestframework==3.13.1
psycopg2-binary==2.9.3